Made some changes that should have fixed a lot of bugs... Still sounds fucked up.
git-svn-id: https://svn.code.sf.net/p/cheesetracker/code@55 88f81a63-fe31-0410-a212-96c38ac4dcd5

+++ cheesetracker/cheesetracker/trackercore/voice_control.h
@@ -67,7 +67,7 @@ struct Voice_Control {
 		Uint32 base_speed;        /* what finetune to use */
 		int period;
 		//sample start index
-        	int offset;        /* The starting byte index in the sample */
+        	size_t offset;        /* The starting byte index in the sample */
 		Sample* ptr;
 		
 		int index; //which samples (if applies, -1 if played from somewhere else )
